backend-events/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]